pyproject.toml
[project]
name = "dasein-as-a-degenerate"
version = "0.1.0"
description = "Bloomberg Terminal for event contract degeneracy"
# hacky removal to appease Railway
# readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "cryptography>=44.0.0",
    "ipykernel>=6.29.5",
    "kalshi-python>=2.0.0",
    "python-dotenv>=0.1.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "pandas>=2.2.3",
    "tqdm>=4.67.1",
    "requests-cache>=1.2.1",
    "discord.py>=2.4.0",
    "watchdog>=6.0.0",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "aio-pika>=9.5.4",
]
# todo, figure out how to setup dev dependencies style workflow

[project.scripts]
start = "src.daad.main:run"
dev = "src.daad.watch:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
